misc/requirements.txt
sqlmodel
python-dotenv
psycopg2-binary